fix neonToken Interface

+++ contracts/precompiled/NeonToken.sol
@@ -4,7 +4,7 @@ pragma solidity >=0.7.6;
 interface INeonWithdraw {
     function withdraw(bytes32) external payable returns (bool);
 
-    function withdraw_on_chain(uint64 chain_id, bytes32 to, uint256 amount) external returns (bool);
+    function withdraw_on_chain(uint256 chain_id, bytes32 to, uint256 amount) external returns (bool);
 
 }
 
@@ -29,7 +29,7 @@ contract NeonToken {
         NeonPrecompiled.withdraw_on_chain(chain_id, spender, msg.value);
     }
 
-    function withdraw_on_chain(uint64 chain_id, bytes32 spender) external payable {
+    function withdraw_on_chain(uint256 chain_id, bytes32 spender) external payable {
         NeonPrecompiled.withdraw_on_chain(chain_id, spender, msg.value);
     }
 }
